added dates detection feature
Added a dates detection feature in the display notes function. That is the ability of a function to detect whether the string of text entered by the user contains any dates and if it does, then it adds all of them to the additional string on the card of the note.